requirements.txt
Django==1.11.2
djangorestframework==3.6.3
Markdown==2.6.8
olefile==0.44
Pillow==4.1.1
pytz==2017.2
